producer/requirements.txt
confluent-kafka==2.8.0
websockets==14.2
python-dotenv==1.1.0